requirements.txt
Flask==3.0.0
SQLAlchemy==2.0.25
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.3
Pillow==10.1.0
pytest==7.4.4
python-jose==3.3.0
python-dateutil==2.8.2
PyYAML==6.0.1
redis==5.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
Werkzeug==3.0.1
Jinja2==3.1.3
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
six==1.16.0